Results page (#51)
* Current events changes

* Refactor fetchCycle to use ResponseCycleType, add fetchEventCycles to API

* Refactor onboarding flow and remove temporary registration status

* Update routes and add fetchEvent function

* Refactor user registration flow and fix navigation

* Add Results page

* Refactor navigation and add results button

This commit refactors the navigation in the App and Header components to use the `Navigate` component from React Router instead of the `navigate` function. It also adds a "Results" button in the Vote component that navigates to the results page. These changes improve the user experience and make the code more readable and maintainable.

* Create Event page

* Remove Results button and add "see results" link

* Fix hex color

* Update Results page with queried statistics

* Add icons

* Add results leaderboard

---------

Co-authored-by: Diego Alzate <diego35.da@gmail.com>